tc: Support larger rate limits
Currently "tc" uses 32bit unsigned integers to represent rate limits in
bytes/second. This restricts maximum rate limit that can be configured
to around 4GiBps = 32gibit. For larger rate limits, "tc" silently fails
and the rate limit is truncated to 32bits thereby configuring incorrect
rate limits.

This patch utilizes the 64bit unsigned int ratespecs in the kernel and
allows configuring much larger rate limits such as for 40gbit Ethernet.

Example with HTB qdisc:

tc qdisc add dev eth1 root handle 1: htb default 1

tc class add dev eth1 classid 1:1 parent 1: htb rate 35000Mbit \
        mtu 65535

tc class show dev eth1

With old 32bit rate limits:
class htb 1:1 root prio 0 rate 640262Kbit ceil 640262Kbit burst 65386b \
        cburst 65386b

With new 64bit rate limits:
class htb 1:1 root prio 0 rate 35000Mbit ceil 35000Mbit burst 56875b \
        cburst 56875b

+++ include/linux/pkt_sched.h
@@ -35,7 +35,7 @@ struct tc_stats {
 	__u32	drops;			/* Packets dropped because of lack of resources */
 	__u32	overlimits;		/* Number of throttle events when this
 					 * flow goes out of allocated bandwidth */
-	__u32	bps;			/* Current flow byte rate */
+	__u64	bps;			/* Current flow byte rate */
 	__u32	pps;			/* Current flow packet rate */
 	__u32	qlen;
 	__u32	backlog;
@@ -79,7 +79,7 @@ struct tc_ratespec {
 	unsigned short	overhead;
 	short		cell_align;
 	unsigned short	mpu;
-	__u32		rate;
+	__u64		rate;
 };
 
 #define TC_RTAB_SIZE	1024
@@ -368,9 +368,9 @@ struct tc_hfsc_qopt {
 };
 
 struct tc_service_curve {
-	__u32	m1;		/* slope of the first segment in bps */
+	__u64	m1;		/* slope of the first segment in bps */
 	__u32	d;		/* x-projection of the first segment in us */
-	__u32	m2;		/* slope of the second segment in bps */
+	__u64	m2;		/* slope of the second segment in bps */
 };
 
 struct tc_hfsc_stats {
@@ -541,7 +541,7 @@ struct tc_netem_corrupt {
 };
 
 struct tc_netem_rate {
-	__u32	rate;	/* byte/s */
+	__u64	rate;	/* byte/s */
 	__s32	packet_overhead;
 	__u32	cell_size;
 	__s32	cell_overhead;
