+++ tc/tc_red.h
@@ -3,6 +3,6 @@
 
 extern int tc_red_eval_P(unsigned qmin, unsigned qmax, double prob);
 extern int tc_red_eval_ewma(unsigned qmin, unsigned burst, unsigned avpkt);
-extern int tc_red_eval_idle_damping(int wlog, unsigned avpkt, unsigned bandwidth, __u8 *sbuf);
+extern int tc_red_eval_idle_damping(int wlog, unsigned avpkt, __u64 bandwidth, __u8 *sbuf);
 
 #endif
